refactor: share Meridian config
Rework spark-api to use the contract configuration from spark-publish.
