--- lib/meridian/abi.json
@@ -1,1 +1,0 @@
-[{"type":"constructor","inputs":[{"internalType":"address","name":"admin","type":"address"}]},{"type":"function","name":"DEFAULT_ADMIN_ROLE","inputs":[],"outputs":[{"internalType":"bytes32","name":"","type":"bytes32"}],"stateMutability":"view"},{"type":"function","name":"EVALUATE_ROLE","inputs":[],"outputs":[{"internalType":"bytes32","name":"","type":"bytes32"}],"stateMutability":"view"},{"type":"function","name":"addMeasurements","inputs":[{"internalType":"string","name":"cid","type":"string"}],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"nonpayable"},{"type":"function","name":"adminAdvanceRound","inputs":[],"outputs":[],"stateMutability":"nonpayable"},{"type":"function","name":"currentRoundIndex","inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view"},{"type":"function","name":"currentRoundMeasurementCount","inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view"},{"type":"function","name":"getRoleAdmin","inputs":[{"internalType":"bytes32","name":"role","type":"bytes32"}],"outputs":[{"internalType":"bytes32","name":"","type":"bytes32"}],"stateMutability":"view"},{"type":"function","name":"getRound","inputs":[{"internalType":"uint256","name":"index","type":"uint256"}],"outputs":[{"internalType":"struct ImpactEvaluator.Round","name":"","type":"tuple","components":[{"type":"uint256"},{"type":"string[]"},{"type":"address[]"},{"type":"uint256[]"},{"type":"bool"},{"type":"string"}]}],"stateMutability":"view"},{"type":"function","name":"grantRole","inputs":[{"internalType":"bytes32","name":"role","type":"bytes32"},{"internalType":"address","name":"account","type":"address"}],"outputs":[],"stateMutability":"nonpayable"},{"type":"function","name":"hasRole","inputs":[{"internalType":"bytes32","name":"role","type":"bytes32"},{"internalType":"address","name":"account","type":"address"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"stateMutability":"view"},{"type":"function","name":"nextRoundLength","inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view"},{"type":"function","name":"renounceRole","inputs":[{"internalType":"bytes32","name":"role","type":"bytes32"},{"internalType":"address","name":"callerConfirmation","type":"address"}],"outputs":[],"stateMutability":"nonpayable"},{"type":"function","name":"revokeRole","inputs":[{"internalType":"bytes32","name":"role","type":"bytes32"},{"internalType":"address","name":"account","type":"address"}],"outputs":[],"stateMutability":"nonpayable"},{"type":"function","name":"roundReward","inputs":[],"outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view"},{"type":"function","name":"rounds","inputs":[{"internalType":"uint256","name":"","type":"uint256"}],"outputs":[{"internalType":"uint256","name":"end","type":"uint256"},{"internalType":"bool","name":"scoresSubmitted","type":"bool"},{"internalType":"string","name":"summaryText","type":"string"}],"stateMutability":"view"},{"type":"function","name":"setNextRoundLength","inputs":[{"internalType":"uint256","name":"_nextRoundLength","type":"uint256"}],"outputs":[],"stateMutability":"nonpayable"},{"type":"function","name":"setRoundReward","inputs":[{"internalType":"uint256","name":"_roundReward","type":"uint256"}],"outputs":[],"stateMutability":"nonpayable"},{"type":"function","name":"setScores","inputs":[{"internalType":"uint256","name":"roundIndex","type":"uint256"},{"internalType":"address payable[]","name":"addresses","type":"address[]"},{"internalType":"uint256[]","name":"scores","type":"uint256[]"},{"internalType":"string","name":"summaryText","type":"string"}],"outputs":[],"stateMutability":"nonpayable"},{"type":"function","name":"supportsInterface","inputs":[{"internalType":"bytes4","name":"interfaceId","type":"bytes4"}],"outputs":[{"internalType":"bool","name":"","type":"bool"}],"stateMutability":"view"},{"type":"event","name":"MeasurementsAdded","inputs":[{"name":"cid","type":"string","indexed":false},{"name":"roundIndex","type":"uint256","indexed":false}],"anonymous":false},{"type":"event","name":"RoleAdminChanged","inputs":[{"name":"role","type":"bytes32","indexed":true},{"name":"previousAdminRole","type":"bytes32","indexed":true},{"name":"newAdminRole","type":"bytes32","indexed":true}],"anonymous":false},{"type":"event","name":"RoleGranted","inputs":[{"name":"role","type":"bytes32","indexed":true},{"name":"account","type":"address","indexed":true},{"name":"sender","type":"address","indexed":true}],"anonymous":false},{"type":"event","name":"RoleRevoked","inputs":[{"name":"role","type":"bytes32","indexed":true},{"name":"account","type":"address","indexed":true},{"name":"sender","type":"address","indexed":true}],"anonymous":false},{"type":"event","name":"RoundStart","inputs":[{"name":"roundIndex","type":"uint256","indexed":false}],"anonymous":false},{"type":"error","name":"AccessControlBadConfirmation","inputs":[]},{"type":"error","name":"AccessControlUnauthorizedAccount","inputs":[{"internalType":"address","name":"account","type":"address"},{"internalType":"bytes32","name":"neededRole","type":"bytes32"}]},{"type":"receive"}]
