+++ lib/round-tracker.js
@@ -1,5 +1,5 @@
 import Sentry from '@sentry/node'
-import { createMeridianContract } from './meridian/contract.js'
+import { createMeridianContract } from './ie-contract.js'
 
 export async function createRoundGetter (pgClient) {
   const contract = await createMeridianContract()
